Create Learning_flow.md
I am a learner who want to dive into the field of Data Science. During COVID 19 pandemic I had attended numerous webinar’s in the field of computer science. As a result of that I made my mind to pursue in the field of AI. I am learning many concepts and codes from scratch and have a dream to become a Data Scientist. 

I’ve learnt that I have still a lot to learn